Integrate point reader and follower, there's still a bizarre bug I can't find

+++ points_module.f90
@@ -4,7 +4,7 @@ module points_module
 
   implicit none
   real, dimension(:,:), allocatable :: points_start, points_move
-  real, dimension(:), allocatable :: points_hit
+  integer, dimension(:), allocatable :: points_hit
   real :: points_dphi
   integer :: n_iter, points_number
